Task/jitpack (#3)
This pull request includes updates to the `README.MD` file and the
addition of a new configuration file. These changes are aimed at
improving documentation and updating the project configuration.

Documentation updates:

*
[`README.MD`](diffhunk://#diff-01e6d9ffed056a02cae8d8a0ec5d476a64d017bf85c0d5a94bb23ca21f33f5aaR35-L40):
Added a new section on how to include the Mattermost Spring Boot Starter
in your project, including the necessary dependency and repository
configurations for `pom.xml`.

Configuration updates:

*
[`jitpack.yml`](diffhunk://#diff-1604193a5efbe3735c6ab8b44601ad6e42df52f6834bcc155650e3eef5b3fa94R1-R2):
Added configuration to specify the use of `openjdk21`.